# docs/swagger.yaml
basePath: /
definitions:
...
  types.APIError:
    properties:
      code:
        type: string
      details:
        type: string
      message:
        type: string
    type: object
  types.APIResponse:
    properties:
      data: {}
      error:
        $ref: '#/definitions/types.APIError'
      success:
        type: boolean
    type: object
...
  types.CreateOrImportTimelockContractRequest:
    properties:
      chain_id:
        type: integer
      contract_address:
        type: string
      is_imported:
        type: boolean
      remark:
        maxLength: 500
        type: string
      standard:
        enum:
        - compound
        - openzeppelin
        type: string
    required:
    - chain_id
    - contract_address
    - standard
    type: object
...
  types.DeleteTimeLockRequest:
    properties:
      chain_id:
        type: integer
      contract_address:
        type: string
      standard:
        enum:
        - compound
        - openzeppelin
        type: string
    required:
    - chain_id
    - contract_address
    - standard
    type: object
...
  types.GetTimeLockDetailRequest:
    properties:
      chain_id:
        type: integer
      contract_address:
        type: string
      standard:
        enum:
        - compound
        - openzeppelin
        type: string
    required:
    - chain_id
    - contract_address
    - standard
    type: object
  types.GetTimeLockDetailResponse:
    properties:
      compound_data:
        $ref: '#/definitions/types.CompoundTimeLockWithPermission'
      openzeppelin_data:
        $ref: '#/definitions/types.OpenzeppelinTimeLockWithPermission'
      standard:
        type: string
    type: object
  types.GetTimeLockListRequest:
    properties:
      standard:
        type: string
      status:
        type: string
    type: object
  types.GetTimeLockListResponse:
    properties:
      compound_timelocks:
        items:
          $ref: '#/definitions/types.CompoundTimeLockWithPermission'
        type: array
      openzeppelin_timelocks:
        items:
          $ref: '#/definitions/types.OpenzeppelinTimeLockWithPermission'
        type: array
      total:
        type: integer
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: TimeLocker Backend API
  title: TimeLocker Backend API
  version: "1.0"
paths:
...
  /api/v1/timelock/create-or-import:
    post:
      consumes:
      - application/json
      description: 创建新的或导入已存在的timelock合约记录。系统会从链上读取合约数据并验证其是否为有效的timelock合约。支持Compound和OpenZeppelin两种标准。合约地址必须为有效以太坊地址（0x
        + 40位十六进制）。
      parameters:
      - description: 创建或导入timelock合约的请求体（地址从鉴权获取）
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/types.CreateOrImportTimelockContractRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 成功创建或导入timelock合约记录
          schema:
            allOf:
            - $ref: '#/definitions/types.APIResponse'
            - properties:
                data:
                  type: object
              type: object
        "400":
          description: 请求参数错误或标准/地址无效（INVALID_STANDARD / INVALID_CONTRACT_ADDRESS）
          schema:
            allOf:
            - $ref: '#/definitions/types.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/types.APIError'
              type: object
        "401":
          description: 未认证或令牌无效
          schema:
            allOf:
            - $ref: '#/definitions/types.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/types.APIError'
              type: object
        "409":
          description: timelock合约已存在
          schema:
            allOf:
            - $ref: '#/definitions/types.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/types.APIError'
              type: object
        "422":
          description: 参数校验失败
          schema:
            allOf:
            - $ref: '#/definitions/types.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/types.APIError'
              type: object
        "500":
          description: 服务器内部错误
          schema:
            allOf:
            - $ref: '#/definitions/types.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/types.APIError'
              type: object
      security:
      - BearerAuth: []
      summary: 创建或导入timelock合约记录
      tags:
      - Timelock
  /api/v1/timelock/delete:
    post:
      consumes:
      - application/json
      description: 硬删除指定的timelock合约记录。只有合约的创建者/导入者才能删除合约记录。删除操作是硬删除，数据从数据库中删除。合约地址必须为有效以太坊地址（0x
        + 40位十六进制）。
      parameters:
      - description: 删除请求体（地址从鉴权获取）
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/types.DeleteTimeLockRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 成功删除timelock合约记录
          schema:
            allOf:
            - $ref: '#/definitions/types.APIResponse'
            - properties:
                data:
                  type: object
              type: object
        "400":
          description: 请求参数错误或标准/地址无效（INVALID_STANDARD / INVALID_CONTRACT_ADDRESS）
          schema:
            allOf:
            - $ref: '#/definitions/types.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/types.APIError'
              type: object
        "401":
          description: 未认证或令牌无效
          schema:
            allOf:
            - $ref: '#/definitions/types.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/types.APIError'
              type: object
        "403":
          description: 无权访问此timelock合约
          schema:
            allOf:
            - $ref: '#/definitions/types.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/types.APIError'
              type: object
        "404":
          description: timelock合约不存在
          schema:
            allOf:
            - $ref: '#/definitions/types.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/types.APIError'
              type: object
        "422":
          description: 参数校验失败
          schema:
            allOf:
            - $ref: '#/definitions/types.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/types.APIError'
              type: object
        "500":
          description: 服务器内部错误
          schema:
            allOf:
            - $ref: '#/definitions/types.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/types.APIError'
              type: object
      security:
      - BearerAuth: []
      summary: 删除timelock合约记录
      tags:
      - Timelock
  /api/v1/timelock/detail:
    post:
      consumes:
      - application/json
      description: 获取指定timelock合约的完整详细信息，包括合约的基本信息、治理参数以及用户权限信息。只有具有相应权限的用户才能查看详细信息。合约地址必须为有效以太坊地址（0x
        + 40位十六进制）。
      parameters:
      - description: 获取详情请求体
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/types.GetTimeLockDetailRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 成功获取timelock合约详情
          schema:
            allOf:
            - $ref: '#/definitions/types.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/types.GetTimeLockDetailResponse'
              type: object
        "400":
          description: 请求参数错误或标准/地址无效（INVALID_STANDARD / INVALID_CONTRACT_ADDRESS）
          schema:
            allOf:
            - $ref: '#/definitions/types.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/types.APIError'
              type: object
        "401":
          description: 未认证或令牌无效
          schema:
            allOf:
            - $ref: '#/definitions/types.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/types.APIError'
              type: object
        "403":
          description: 无权访问此timelock合约
          schema:
            allOf:
            - $ref: '#/definitions/types.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/types.APIError'
              type: object
        "404":
          description: timelock合约不存在
          schema:
            allOf:
            - $ref: '#/definitions/types.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/types.APIError'
              type: object
        "422":
          description: 参数校验失败
          schema:
            allOf:
            - $ref: '#/definitions/types.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/types.APIError'
              type: object
        "500":
          description: 服务器内部错误
          schema:
            allOf:
            - $ref: '#/definitions/types.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/types.APIError'
              type: object
      security:
      - BearerAuth: []
      summary: 获取timelock合约详细信息
      tags:
      - Timelock
  /api/v1/timelock/list:
    post:
      consumes:
      - application/json
      description: 获取当前用户在所有链上有权限访问的timelock合约列表。支持按合约标准和状态进行筛选。返回的列表根据用户权限进行精细控制，只显示用户作为创建者、管理员、提议者、执行者的合约。
      parameters:
      - description: 查询参数
        in: body
        name: request
        schema:
          $ref: '#/definitions/types.GetTimeLockListRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 成功获取timelock合约列表
          schema:
            allOf:
            - $ref: '#/definitions/types.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/types.GetTimeLockListResponse'
              type: object
        "400":
          description: 请求参数错误或标准无效
          schema:
            allOf:
            - $ref: '#/definitions/types.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/types.APIError'
              type: object
        "401":
          description: 未认证或令牌无效
          schema:
            allOf:
            - $ref: '#/definitions/types.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/types.APIError'
              type: object
        "500":
          description: 服务器内部错误
          schema:
            allOf:
            - $ref: '#/definitions/types.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/types.APIError'
              type: object
      security:
      - BearerAuth: []
      summary: 获取用户timelock合约列表（按权限筛选，所有链）
      tags:
      - Timelock
  /api/v1/timelock/refresh-permissions:
    post:
      consumes:
      - application/json
      description: 刷新该用户在库中所有timelock合约的权限，将合约中所有角色获取一遍然后更新数据库。这个操作会从区块链上重新读取用户在各个timelock合约中的最新权限信息。
      produces:
      - application/json
      responses:
        "200":
          description: 成功刷新权限
          schema:
            allOf:
            - $ref: '#/definitions/types.APIResponse'
            - properties:
                data:
                  type: object
              type: object
        "401":
          description: 未认证或令牌无效
          schema:
            allOf:
            - $ref: '#/definitions/types.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/types.APIError'
              type: object
        "500":
          description: 服务器内部错误
          schema:
            allOf:
            - $ref: '#/definitions/types.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/types.APIError'
              type: object
      security:
      - BearerAuth: []
      summary: 刷新用户所有timelock合约权限
      tags:
      - Timelock
  /api/v1/timelock/update:
    post:
      consumes:
      - application/json
      description: 更新指定timelock合约的备注信息。只有合约的创建者/导入者才能更新备注。备注信息用于帮助用户管理和识别不同的timelock合约。合约地址必须为有效以太坊地址（0x
        + 40位十六进制）。
      parameters:
      - description: 更新请求体（地址从鉴权获取）
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/types.UpdateTimeLockRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 成功更新timelock合约备注
          schema:
            allOf:
            - $ref: '#/definitions/types.APIResponse'
            - properties:
                data:
                  type: object
              type: object
        "400":
          description: 请求参数错误或标准/地址无效（INVALID_STANDARD / INVALID_CONTRACT_ADDRESS）
          schema:
            allOf:
            - $ref: '#/definitions/types.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/types.APIError'
              type: object
        "401":
          description: 未认证或令牌无效
          schema:
            allOf:
            - $ref: '#/definitions/types.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/types.APIError'
              type: object
        "403":
          description: 无权访问此timelock合约
          schema:
            allOf:
            - $ref: '#/definitions/types.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/types.APIError'
              type: object
        "404":
          description: timelock合约不存在
          schema:
            allOf:
            - $ref: '#/definitions/types.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/types.APIError'
              type: object
        "422":
          description: 参数校验失败
          schema:
            allOf:
            - $ref: '#/definitions/types.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/types.APIError'
              type: object
        "500":
          description: 服务器内部错误
          schema:
            allOf:
            - $ref: '#/definitions/types.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/types.APIError'
              type: object
      security:
      - BearerAuth: []
      summary: 更新timelock合约备注
      tags:
      - Timelock
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"